pyproject.toml
[tool.poetry]
name = "system-design-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
# packages = [{include = "system-design-project"}]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
todoist-api-python = "^2.0.2"
black = "^22.10.0"
isort = "^5.10.1"
pycodestyle = "^2.9.1"
django = "^4.1.3"
loguru = "^0.6.0"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
